+++ libs/rcore/rcolor.h
@@ -5,8 +5,8 @@
 
 #include "../conf_raylib.h"
 
-Color TRNSP(Color color, float transparent);
+Color TRNSP(Color color, double transparent);
 
-Color TRANSCOLOR(const Color &source, const Color &destination, float percent);
+Color TRANSCOLOR(const Color &source, const Color &destination, double percent);
 
 #endif
